% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_probs.R
\name{aerosol_emission}
\alias{aerosol_emission}
\title{Aerosol Emission per Hour}
\usage{
aerosol_emission(
  emission_breathing = 0.06,
  emission_speaking = 0.6,
  speaking_ratio = 0.1,
  resp_rate = 10
)
}
\arguments{
\item{emission_breathing}{a double, emission while breathing cm^3}

\item{emission_speaking}{a double, emission while speaking cm^3}

\item{speaking_ratio}{a double, between 0 and 1 for ratio between speaking and breathing}

\item{resp_rate}{a double, the respiratory rate in liters per minute}
}
\description{
Estimate the aerosol emission rate per hours
}
