% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_risk.R
\name{calc_risk}
\alias{calc_risk}
\title{Converts Infected persons, Number of Contact and Population
to the associated probability of a contact}
\usage{
calc_risk(I, n, pop)
}
\arguments{
\item{I}{integer, the number of persons infected}

\item{n}{integer, the number of contacts per day}

\item{pop}{integer, the total population}
}
\description{
Uses a Taylor approximation for repeated Binomial trials
Classically associated with the "Birthday Problem"
}
\details{
P(A) = 1- P(A)'
}
