% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_community_immunity.R
\name{calculate_community_immunity}
\alias{calculate_community_immunity}
\title{Calculating Community Immunity}
\source{
Lopman, B. A. et al. A framework for monitoring population immunity 
    to SARS-CoV-2. Annals of Epidemiology S1047279721002635 (2021) 
    \url{doi:10.1016/j.annepidem.2021.08.013}.
}
\usage{
calculate_community_immunity(prop_infected, prop_vaccinated)
}
\arguments{
\item{prop_infected}{a numeric representing the proportion of the population
with natural infection}

\item{prop_vaccinated}{a numeric representing the proportion of the
population that is vaccinated.}
}
\description{
Considering Proportion of the Vaccinated and Those Who Were Naturall
Infected
}
\examples{
{

calculate_community_immunity(.3,.5)

plot(seq(.4,.7,.1),
calculate_community_immunity(prop_infected = seq(.4,.7,.1), 
														 prop_vaccinated = .5), 
														 ylab = "Community Immunity", 
														 xlab = "Proportion Infected",
														 main = "Community Immunity Given 50\% Vaccinated", 
														 adj =0)

}
}
