% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_transmission_increase.R
\name{calculate_transmission_increase}
\alias{calculate_transmission_increase}
\title{Transmission Increase}
\usage{
calculate_transmission_increase(R_w, rho, D = 4.7, omega = 0)
}
\arguments{
\item{R_w}{a numeric, the baseline effective reproduction number}

\item{rho}{a numeric, the estimated transmission advantage}

\item{D}{a numeric, the generation time (average)}

\item{omega}{a numeric with default of NULL indicating proportion of the 
population that has immunity against earlier variants}
}
\description{
Calculating the Transmission Increase from Factors
}
\examples{
c(4,6) * calculate_transmission_increase(R_w = .8, .43, D = 5.2, omega = NULL)

c(4,6) * calculate_transmission_increase(R_w = .8, .43, D = 5.2, omega = .25)

}
