% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_mortality_data}
\alias{covid_mortality_data}
\title{Mortality Estimates from Verity Et Al for Covid}
\format{
data frame
}
\usage{
covid_mortality_data
}
\description{
Mortality Estimates from Verity Et Al for Covid
}
\keyword{datasets}
