% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positivity_adjustment.R
\name{decrease_cases}
\alias{decrease_cases}
\title{Decrease Cases}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4809590/}
}
\usage{
decrease_cases(calculated_cases, pos_rate, m = 10.83, k = 0.5)
}
\arguments{
\item{calculated_cases}{numeric, the calculated cases}

\item{pos_rate}{numeric, the observed positivity rate}

\item{m}{scaling factor}

\item{k}{scaling factor with default of 1/2 on range [0,1]}
}
\description{
Decreased previous observed cases back to scale
}
