% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_ve_wane.R
\name{effective_ve_wane}
\alias{effective_ve_wane}
\title{Vaccine Effectiveness}
\usage{
effective_ve_wane(VE = 0.8, omega = 0.1, t, theta = 0.3, t_2 = 14)
}
\arguments{
\item{VE}{a numeric, the maximum vaccine efficacy}

\item{omega}{a numeric, the rate parameter of wane}

\item{t}{an integer, the day of interest}

\item{theta}{a numeric, the general wane of the vaccine to be calibrated}

\item{t_2}{an integer, the inflection of the heavyside}
}
\description{
Calculates Effective Vaccine Effectiveness as a function of parameters
and time since vaccination. Includes heavyside function to accomendate
the daily from injection
}
\examples{
plot(0:30, effective_ve_wane(t = 0:30))
}
