% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{epi_convolve}
\alias{epi_convolve}
\title{Convolution}
\usage{
epi_convolve(cases, pdf, direction = "forward")
}
\arguments{
\item{cases}{numeric, the observed cases}

\item{pdf}{numeric vector, the pdf of cases}

\item{direction}{string, one of forward or backwards convolution}
}
\description{
Perform a convolution given a pdf
}
\examples{
{
  
true_cases <- rpois(100, 5)
out <- data.frame(reported = ceiling(epi_convolve(true_cases, pdf = c(.2,.2,.3,.2,.1))),
true_cases,
date = 1:100
)

library(ggplot2)
ggplot(out, aes(date))+
	geom_line(aes(y = true_cases))+
	geom_line(aes(y = reported), color = "blue")
}
}
