% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_finalsize.R
\name{epi_finalsize}
\alias{epi_finalsize}
\title{Final Size of An Epidemic}
\usage{
epi_finalsize(r0)
}
\arguments{
\item{r0}{a double, the reproduction number (can be basic or Reff)}
}
\description{
Calculate the final size of an epidemic. 
Usethe the following logic:
}
\details{
1 + 1/R0*W(-R0exp(-R0))
}
\examples{
{
epi_finalsize(3.5)

plot(seq(1,3,.01), epi_finalsize(seq(1,3,.01)), 
   xlab = "R", ylab = "Percent Infected", type = "l")
}

}
