% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_vax_coverage_age.R
\name{generate_vax_coverage_age}
\alias{generate_vax_coverage_age}
\title{Vaccine Coverage by Age}
\usage{
generate_vax_coverage_age(county_use = nccovid::cone_region)
}
\arguments{
\item{county_use}{a string vector representing the North Carolina
counties to be displayed}
}
\description{
Generates a graphic that shows the vaccine coverage by age band as a 
stacked bar chart in order to show the coverage as well as the raw number
of vaccinated versus unvaccinated individuals.
}
