% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cdc_detail.R
\name{get_cdc_detail}
\alias{get_cdc_detail}
\title{Retrieve CDC COVID-19 Spread by County}
\source{
<https://covid.cdc.gov/covid-data-tracker/#datatracker-home>
}
\usage{
get_cdc_detail()
}
\value{
a data.table with details regarding disease spread by county
}
\description{
Utilizes the CDC's COVID-19 Tracker to General Information About the Spread
of Disease.
}
