% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cdc_vax.R
\name{get_cdc_vax}
\alias{get_cdc_vax}
\title{Retrieve CDC Vaccination Rates by County}
\source{
<https://covid.cdc.gov/covid-data-tracker/#datatracker-home>
}
\usage{
get_cdc_vax()
}
\value{
a data.table with vaccination information by county by day
}
\description{
Utilizes the CDC's COVID-19 Tracker to Retrieve Vaccination Rates
}
