% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covid_county_demographics.R
\name{get_county_covid_demographics}
\alias{get_county_covid_demographics}
\title{Get Reported Demographics by County}
\usage{
get_county_covid_demographics(
  demographic = "age_group",
  region = NULL,
  add_population = TRUE
)
}
\arguments{
\item{demographic}{a string, one of age, k_12, ethnicity, gender, or race}

\item{region}{a string vector, could a string or vector of the counties of 
interest}

\item{add_population}{a boolean, TRUE to add the population estimates
(may not be available for all demographics)}
}
\description{
Retrieve Latest Demographic data from NC DHHS.
}
\examples{
library(ggplot2)
library(data.table)
out <- get_county_covid_demographics(region = "Guilford", demographic = "age")

out[,per_capita:=cases/(population/100000)]
ggplot(out, aes(week_of, per_capita, colour = metric))+
geom_line()+
theme_bw()+
theme(legend.position = "top")+
labs(
title = "SARS-CoV-2 Cases per 100k Residents",
subtitle = "Guilford County, North Carolina",
x = "Cases per 100k",
y = NULL
)

}
