% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_county_tests.R
\name{get_county_tests}
\alias{get_county_tests}
\title{Pull Testing Information at County Level}
\usage{
get_county_tests(county_select = NULL)
}
\arguments{
\item{county_select}{a string, the county of counties of interest}
}
\description{
Pulling information on tests and percent positive at the 
county level. Notes that these data are updated manually.
}
