% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covid_county_plus.R
\name{get_covid_county_plus}
\alias{get_covid_county_plus}
\title{Append Additional Information to COVID-19 Count Information}
\usage{
get_covid_county_plus(county = NULL, adjusted = TRUE)
}
\arguments{
\item{county}{a string vector, representing the counties of interest. If left
NULL then all counties in North Carolina will be returned}

\item{adjusted}{a logical, should the adjusted cases be used (to account for
reporting gaps on weekends and holidays from NCDHHS). Defaults to \code{TRUE}}
}
\value{
a data.table with combined values for all counties or specified counties
}
\description{
Combined NCDHHS Case and Death Counts with CDC Details
}
