% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covid_demographics.R
\name{get_covid_demographics}
\alias{get_covid_demographics}
\title{Pull Covid-19 Demographics}
\usage{
get_covid_demographics()
}
\description{
These data originate from the North Carolina DHHS Covid-19 Dashboard
}
