% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covid_state.R
\name{get_covid_state}
\alias{get_covid_state}
\title{Retrieve Covid19 Data}
\usage{
get_covid_state(state = "North Carolina", county = NULL)
}
\arguments{
\item{state}{with a default of North Carolina}

\item{county}{the county, if desired}
}
\description{
This functions hits the Johns Hopkins Covid Data Respository
and filters down to North Carolina Data
}
\examples{
\dontrun{
get_covid_state()
}


}
