% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_mobility.R
\name{get_google_mobility}
\alias{get_google_mobility}
\title{Get Google Mobility Data}
\source{
<https://www.gstatic.com/covid19/mobility/Global_Mobility_Report.csv>
}
\usage{
get_google_mobility(counties = NULL)
}
\arguments{
\item{counties}{a string vector of the desired counties deafult of \code{NULL}
returns all counties.}
}
\description{
This is a helper function to automatically access Google's 
Mobility data for North Carolina Counties
}
\examples{
\dontrun{
# Return All NC Counties
get_google_mobility()

# Return a Specific County
get_google_mobility("Alamance County")

}
}
