% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_mobility2.R
\name{get_google_mobility2}
\alias{get_google_mobility2}
\title{Get Google Mobility (Version 2)
This is a helper function to automatically access Google's 
Mobility data for North Carolina Counties. It's a bit faster than
`get_google_mobility` which downloads the entire csv for the world}
\usage{
get_google_mobility2(counties = NULL, state_in = "North Carolina")
}
\arguments{
\item{counties}{a string vector of the desired counties default of \code{NULL}
returns all counties.}

\item{state_in}{a string vector representing state or states of interest}
}
\description{
Get Google Mobility (Version 2)
This is a helper function to automatically access Google's 
Mobility data for North Carolina Counties. It's a bit faster than
`get_google_mobility` which downloads the entire csv for the world
}
