% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hospitalizations.R
\name{get_hospitalizations}
\alias{get_hospitalizations}
\title{Get Reported Hospitalisation Data}
\usage{
get_hospitalizations()
}
\description{
Retrieve Latest Hospitalisation data from NC DHHS. This is not an automatic
Scrap so there could be some delay.
}
