% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaccinations.R
\name{get_vaccinations}
\alias{get_vaccinations}
\title{Pull Vaccination Rates by County}
\usage{
get_vaccinations(county_list = NULL, add_population = TRUE)
}
\arguments{
\item{county_list}{a string, the county to pull. default is all counties}

\item{add_population}{a boolean, default of TRUE to add a population}
}
\description{
Pulled from <conedatascience/covid-data>
}
\examples{
{

get_vaccinations(county_list = "Guilford")

}
}
