% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positivity_adjustment.R
\name{increase_cases}
\alias{increase_cases}
\title{Increase Cases}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4809590/}
}
\usage{
increase_cases(observed_cases, pos_rate, m = 10.83, k = 0.5)
}
\arguments{
\item{observed_cases}{numberic, the reported cases}

\item{pos_rate}{numberic, the observed positivity rate}

\item{m}{scaling factor}

\item{k}{scaling factor with default of 1/2 on range [0,1]}
}
\description{
Increase case count in the presence of high testing positivity
}
