% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_probs.R
\name{infection_probability}
\alias{infection_probability}
\title{Infection Probability}
\usage{
infection_probability(D50 = 316)
}
\arguments{
\item{D50}{a double, #RNA for 50 percent infection probability (D50)}
}
\description{
Calculate Infection probability given D50
1-10^(log10(.5)/D50)
}
