% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nc_county_demos}
\alias{nc_county_demos}
\title{NC Biological Sex Information by County}
\format{
a data.frame, with NC DHHS county level demographic metrics
}
\source{
American Community Survey (2020 Vintage)
}
\usage{
nc_county_demos
}
\description{
NC Biological Sex Information by County
}
\keyword{datasets}
