% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nc_county_fips}
\alias{nc_county_fips}
\title{NC Fips by County}
\format{
a data.frame
}
\source{
https://www.lib.ncsu.edu/gis/countyfips
}
\usage{
nc_county_fips
}
\description{
NC Fips by County
}
\keyword{datasets}
