% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nc_delay}
\alias{nc_delay}
\title{NCDHHS Reported Delay Distribution}
\format{
a list, with parameters for a gamma distribution fit to delays
}
\source{
https://covid19.ncdhhs.gov/dashboard/testing
}
\usage{
nc_delay
}
\description{
Reported testing and reporting delays were fit using the EpiNow2 
boostrapped gamma fit. The reported delays were manually collected from
the NCDHHS website.
}
\keyword{datasets}
