% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nc_events}
\alias{nc_events}
\title{Key North Carolina Events}
\format{
a data.frame
}
\source{
various and https://www.nc.gov/covid-19/staying-ahead-curve
}
\usage{
nc_events
}
\description{
Key North Carolina Events
}
\keyword{datasets}
