% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nc_fatality_estimates}
\alias{nc_fatality_estimates}
\title{Mortality and Hospitalization Estimates}
\format{data.frame}
\usage{
nc_fatality_estimates
}
\description{
Mortality and Hospitalization Estimates
}
\keyword{datasets}
