% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nc_hc_coalitions}
\alias{nc_hc_coalitions}
\title{NC Hospital Preparedness Coalitions}
\format{
a data.frame, with NC county name and associated coalition
}
\source{
https://easternhpc.com/state-coalition-map/
}
\usage{
nc_hc_coalitions
}
\description{
NC Hospital Preparedness Coalitions
}
\keyword{datasets}
