% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nc_pop_dhhs}
\alias{nc_pop_dhhs}
\title{NC Population by NCDHHS Age Band}
\format{
a data.frame
}
\source{
2020 ACS Survey
}
\usage{
nc_pop_dhhs
}
\description{
NC Population by NCDHHS Age Band
}
\keyword{datasets}
