% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nc_population}
\alias{nc_population}
\title{North Carolina Population by County}
\format{
a data.frame
}
\source{
Source: North Carolina OSBM, Standard Population Estimates, 
    Vintage 2018 and Population Projections, Vintage 2019
    <https://www.osbm.nc.gov/demog/county-estimates>
}
\usage{
nc_population
}
\description{
North Carolina Population by County
}
\keyword{datasets}
