% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nc_svi_county}
\alias{nc_svi_county}
\title{CDC Social Vulnerability for North Carolina by County}
\format{
a data.frame
}
\source{
[U.S. Centers for Disease Prevention and Control (CDC) Social Vulnerability Index (SVI)](https://svi.cdc.gov/)
}
\usage{
nc_svi_county
}
\description{
CDC Social Vulnerability for North Carolina by County
}
\keyword{datasets}
