% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nccovid-package.R
\docType{package}
\name{nccovid-package}
\alias{nccovid}
\alias{nccovid-package}
\title{nccovid: Pull North Carolina Covid-19 Outbreak Information}
\description{
This package allows users to easily retrieve data that have been pulled from the North Carolina Department of Health and Human Services Covid-19 Dashboard.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/conedatascience/nccovid}
  \item \url{https://conedatascience.github.io/nccovid/}
  \item Report bugs at \url{https://github.com/conedatascience/nccovid/issues}
}

}
\author{
\strong{Maintainer}: Michael DeWitt \email{me.dewitt.jr@gmail.com} (\href{https://orcid.org/0000-0001-8940-1967}{ORCID})

Authors:
\itemize{
  \item Jennifer Wenner
  \item  Cone Health [copyright holder]
}

}
\keyword{internal}
