% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_estimates.R
\name{pull_estimates}
\alias{pull_estimates}
\title{Pull Latest Estimates}
\usage{
pull_estimates(region = NULL, variable = "R")
}
\arguments{
\item{region}{a string or string vector, the region or county of interest}

\item{variable}{a string or string vector, the measure of interest}
}
\description{
All estimates for North Carolina and some select regions have been
precalculated and available for download. These estimates are utilising
the EpiNow2 package
}
