% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_hospitalization_demos.R
\name{pull_hospitalization_demos}
\alias{pull_hospitalization_demos}
\title{Pull Hospitalization Demographics}
\usage{
pull_hospitalization_demos()
}
\value{
a data.table
}
\description{
A Helper Function That Pull Hospitalizations by Demographics
as Posted by NCDHHS
}
