% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_vaccine_census.R
\name{pull_vaccine_census}
\alias{pull_vaccine_census}
\title{Pull Vaccine Information by Census Tract}
\usage{
pull_vaccine_census(county_pull = NULL)
}
\arguments{
\item{county_pull}{a vector, the counties to select for the query}
}
\value{
a data.table with vaccination rates at the census tract level
}
\description{
The purpose of this function is to pull data provided by NCDHHS
on vaccination rates by census tract. Additionally, information
on SVI and other demographics is provided
}
