% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2R.R
\name{r2R}
\alias{r2R}
\title{Convert Growth Rate to Reproduction Number}
\source{
\url{https://royalsocietypublishing.org/doi/10.1098/rsif.2020.0144}
}
\usage{
r2R(r, mu_gamma = 4.7, sd_gamma = 2.9)
}
\arguments{
\item{r}{a double representing the growth rate}

\item{mu_gamma}{a double representing mean of the 
gamma function characterizing the generation time.}

\item{sd_gamma}{a double representing standard deviation of the 
gamma function characterizing the generation time.}
}
\description{
Convenience function to convert a growth rate (little r)
to a reproduction number (R) given a particular
generation time.
}
\examples{

cases <- rpois(10,2.5)
t <- 1:10
fit <- glm(cases ~ t, family = "poisson")

r2R(coef(fit)[2])



}
