% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_probs.R
\name{rna_content}
\alias{rna_content}
\title{RNA Content in Aerosol}
\usage{
rna_content(resp_fluid = 5e+08, wet_aerosol_diameter = 5)
}
\arguments{
\item{resp_fluid}{a double, respiratory fluid RNA conc [/cm^3]}

\item{wet_aerosol_diameter}{a double, mean wet aerosol diameter [um]}
}
\description{
Estimate the RNA Content in an Aerosol
}
