% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_shannon.R
\name{calculate_shannon}
\alias{calculate_shannon}
\title{Calculate Shannon Index for a Series}
\usage{
calculate_shannon(incidence)
}
\arguments{
\item{incidence}{numeric, the case incidence}
}
\value{
double, the intensity of epidemic calculated via Shannon index
}
\description{
Calculate Shannon Index for a Series
}
\examples{
dat <- nccovid::get_covid_state(c("Guilford", "Forsyth", "Mecklenburg", "Wake"))

library(dplyr)
library(ggplot2)

counties <- unique(dat$county)
entropy_values = dat \%>\% 
	group_by(county) \%>\% 
	dplyr::group_split() \%>\% 
	lapply( function(x) calculate_shannon(incidence = x$cases_daily) ) \%>\%
	unlist()
	
entropy_values = data.frame(county = counties, intensity = entropy_values)
	entropy_values \%>\% 
	filter(county \%in\% nccovid::triad_counties) \%>\% 
	ggplot(aes(reorder(county,intensity), intensity))+
	geom_point()+
	coord_flip()+
	labs(
		title = "Epidemic Intensity"
	)

}
