% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_shannon.R
\name{calculate_shannon}
\alias{calculate_shannon}
\title{Calculate Shannon Index for a Series}
\usage{
calculate_shannon(incidence)
}
\arguments{
\item{incidence}{numeric, the case incidence}
}
\value{
double, the intensity of epidemic calculated via Shannon index
}
\description{
Calculate Shannon Index for a Series
}
