% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covid_state.R
\name{get_covid_state}
\alias{get_covid_state}
\title{Retrieve Covid19 Data}
\usage{
get_covid_state(
  state = "North Carolina",
  select_county = NULL,
  data_source = c("cone", "hopkins"),
  reporting_adj = FALSE
)
}
\arguments{
\item{state}{with a default of North Carolina}

\item{select_county}{the county, if desired}

\item{data_source}{which data source you would like to use
one of "cone" or "hopkins"}

\item{reporting_adj}{a boolean, default of \code{FALSE} which adjust
for two known issues with North Carolina Report occurring on 
2020-09-25 when all antigen tests were added and on 2020-11-13
where only 10 hours of data were reported and the remaining cases
were rolled into 2020-11-14. Also corrects for reporting lapses on 
Thanksgiving and Christmas Eve/ Christmas Day.}
}
\description{
This functions hits the Johns Hopkins Covid Data Respository
and filters down to North Carolina Data. Alternatively, it can
retrieve any state.
}
\examples{
\dontrun{
# To get all counties
get_covid_state()

# To get a single county
get_covid_state(select_county = "Guilford")
}


}
