% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaccinations_demo.R
\name{get_vaccinations_demo}
\alias{get_vaccinations_demo}
\title{Pull Vaccinations by County and Demographic}
\usage{
get_vaccinations_demo(county_list = NULL, demographic = NULL, status = NULL)
}
\arguments{
\item{county_list}{a string, the county to pull. default is all counties.}

\item{demographic}{a string, the demographic requested. one or more of "race", 
"age", "ethnicity", "sex". default is all demographics}

\item{status}{string, either 'partial' or 'full' for number of people with
partially or fully vaccinated. Default of NULL returns both.}
}
\description{
Pulled from <conedatascience/covid-data>. The column vax_n returns
vaccination counts. Population references are available for the user to 
calculate distribution and/or per capita rates.
}
\examples{
{

get_vaccinations_demo(county_list = "Guilford",
                      demographic = 'race', status = 'partial', latest = TRUE)

}
}
