% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nc_pop_age}
\alias{nc_pop_age}
\title{North Carolina Population by County by Age}
\format{
a data.frame
}
\source{
ACS2020 Metrics
}
\usage{
nc_pop_age
}
\description{
North Carolina Population by County by Age
}
\keyword{datasets}
