% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nccovid-package.R
\docType{package}
\name{nccovid-package}
\alias{nccovid}
\alias{nccovid-package}
\title{nccovid: Pull North Carolina Covid-19 Outbreak Information}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael DeWitt \email{me.dewitt.jr@gmail.com} (\href{https://orcid.org/0000-0001-8940-1967}{ORCID})

}
\keyword{internal}
